requirements.txt
apturl==0.5.2
attrs==21.2.0
bcrypt==3.1.7
blinker==1.4
botocore==1.21.22
Brlapi==0.7.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
chrome-gnome-shell==0.0.0
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
cycler==0.10.0
dbus-python==1.2.16
decorator==5.0.9
defer==1.0.6
distro==1.4.0
distro-info===0.23ubuntu1
dj-database-url==0.5.0
Django==2.2.12
django-ckeditor==6.1.0
django-cleanup==5.1.0
django-heroku==0.3.1
django-js-asset==1.2.2
django-taggit==1.5.1
docker==4.4.1
entrypoints==0.3
fasteners==0.14.1
future==0.18.2
google-api-core==1.31.2
google-api-python-client==1.7.11
google-auth==1.5.1
google-auth-httplib2==0.0.3
googleapis-common-protos==1.53.0
gunicorn==20.0.4
httplib2==0.14.0
idna==2.8
jmespath==0.10.0
jsonschema==3.2.0
keyring==18.0.1
kiwisolver==1.3.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
MarkupSafe==1.1.0
matplotlib==3.4.2
monotonic==1.5
netifaces==0.10.4
numpy==1.21.1
oauth2client==4.1.2
oauthlib==3.1.0
olefile==0.46
packaging==21.0
pexpect==4.6.0
Pillow==7.0.0
protobuf==3.6.1
psycopg2==2.8.6
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycups==1.9.73
PyDrive==1.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyOpenSSL==19.0.0
pyparsing==2.4.7
pyRFC3339==1.1
pyrsistent==0.18.0
python-apt==2.0.0+ubuntu0.20.4.5
python-dateutil==2.7.3
python-debian===0.1.36ubuntu1
python-dotenv==0.19.0
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.22.0
requests-unixsocket==0.2.0
rsa==4.0
s3transfer==0.5.0
screen-resolution-extra==0.0.0
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
SQLAlchemy==1.3.12
sqlparse==0.2.4
style==1.1.0
systemd-python==234
ubuntu-advantage-tools==27.2
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
update==0.0.1
uritemplate==0.6
urllib3==1.25.8
wadllib==1.3.3
websocket-client==0.57.0
whitenoise==5.2.0
xkit==0.0.0
